requirements.txt
pyaudio>=0.2.13
numpy>=1.24.0
openai-whisper>=20231117
requests>=2.31.0
pyperclip>=1.8.2
rumps>=0.4.0
pyyaml>=6.0.1
